requirements - 복사본.txt
requests
python-dotenv
pandas
streamlit
